Delete labels (#101)
* Initial work on deleting labels

Implemented initial label deletion support in the UI. Also fixed an issue that prevented users from creating multiple labels in a row (the form would silently never submit).

* Confirm label deletion

* Remove debug code

+++ pages/app/workspace/[id]/settings/labels.vue
@@ -1,17 +1,18 @@
 <script setup lang="ts">
-import { WORKSPACE_DATA_KEY } from "~/lib/injection-keys";
-import {
-  FormItem,
-  FormField,
-  FormLabel,
-  FormControl,
-} from "~/components/ui/form";
-import { useForm } from "vee-validate";
-import { createWorkspaceLabelSchema } from "~/lib/validation";
-import { toTypedSchema } from "@vee-validate/zod";
 import { Cross2Icon, PlusIcon } from "@radix-icons/vue";
-import type { Label } from "~/server/db/schema";
+import { toTypedSchema } from "@vee-validate/zod";
+import { useForm } from "vee-validate";
+import { toast } from "vue-sonner";
 import EasyTooltip from "~/components/easy-tooltip.vue";
+import {
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+} from "~/components/ui/form";
+import { WORKSPACE_DATA_KEY } from "~/lib/injection-keys";
+import { createWorkspaceLabelSchema } from "~/lib/validation";
+import type { Label } from "~/server/db/schema";
 
 const { workspace, labels } = inject(WORKSPACE_DATA_KEY)!;
 
@@ -28,18 +29,36 @@ const form = useForm({
     name: "",
     color: "#00FF00",
   },
+  keepValuesOnUnmount: true,
 });
 
 const { mutateAsync } = useCreateLabel();
+const { mutateAsync: deleteLabelMutation } = useDeleteLabel();
 
-const onSubmit = form.handleSubmit(async (values) => {
+const onSubmit = form.handleSubmit(async (values, { resetForm }) => {
   await mutateAsync(values);
-  form.resetForm();
-  router.push({ query: {} });
+
+  resetForm({
+    values: {
+      workspaceId: workspace.value?.id ?? "",
+      name: "",
+      color: "#00FF00",
+    },
+  });
+
+  router.replace({ query: {} });
 });
 
 const deleteLabel = async (_label: Label) => {
-  // TODO: Implement delete label
+  try {
+    await deleteLabelMutation({
+      labelId: _label.id,
+      workspaceId: _label.workspaceId,
+    });
+    toast.success("Label deleted successfully.");
+  } catch {
+    toast.error("Failed to delete label.");
+  }
 };
 </script>
 
@@ -74,16 +93,38 @@ const deleteLabel = async (_label: Label) => {
           :style="{ backgroundColor: label.color }"
         />
         <span>{{ label.name }}</span>
-        <EasyTooltip tooltip="Delete label">
-          <Button
-            size="icon"
-            variant="ghost"
-            class="hidden p-0 w-auto h-auto ml-auto"
-            @click="() => deleteLabel(label)"
-          >
-            <Cross2Icon class="w-4 h-4" />
-          </Button>
-        </EasyTooltip>
+        <AlertDialog>
+          <EasyTooltip tooltip="Delete label">
+            <AlertDialogTrigger as-child>
+              <Button
+                size="icon"
+                variant="ghost"
+                class="hidden p-1 w-auto h-auto ml-auto"
+              >
+                <Cross2Icon class="w-4 h-4" />
+              </Button>
+            </AlertDialogTrigger>
+          </EasyTooltip>
+
+          <AlertDialogContent>
+            <AlertDialogHeader>
+              <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+              <AlertDialogDescription>
+                This action cannot be undone. This will permanently delete the
+                label and remove it from all tasks.
+              </AlertDialogDescription>
+            </AlertDialogHeader>
+            <AlertDialogFooter>
+              <AlertDialogCancel>Cancel</AlertDialogCancel>
+              <AlertDialogAction
+                variant="destructive"
+                @click="() => deleteLabel(label)"
+              >
+                Delete
+              </AlertDialogAction>
+            </AlertDialogFooter>
+          </AlertDialogContent>
+        </AlertDialog>
       </div>
     </div>
   </div>
@@ -97,7 +138,7 @@ const deleteLabel = async (_label: Label) => {
       <FormField v-slot="{ componentField }" name="workspaceId">
         <FormItem>
           <FormControl>
-            <input type="hidden" v-bind="componentField" >
+            <input type="hidden" v-bind="componentField" />
           </FormControl>
         </FormItem>
       </FormField>
